google-adk/example/pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
description = "Example hello world project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hypercorn>=0.17.3",
    "pydantic>=2.10.6",
    "restate_sdk[serde]>=0.11.0",
    "google-adk>=1.17.0",
    "litellm"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
